Initial connect ledger code restructure
